server/requirements.txt
selenium
beautifulsoup4
psycopg2
python-dotenv
undetected-chromedriver==3.1.0
Django==3.2.12
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
django-cors-headers==3.7.0